requirements.txt
mistralai>=0.4.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-mistralai>=0.0.4
colorama>=0.4.6
pydantic>=2.0.0